==> trash/04702024/01.0_Preprocessing of Environment.ipynb <==
## 

### Preprocess single cell data with cellranger

!source /home2/lumpi/.bashrc

!export PATH="/home2/lumpi/miniconda3/bin:$PATH"

!echo $PATH

!conda info

!/home2/lumpi/miniconda3/bin/conda info

### cellranger to form feature counts
cellranger count --id=sample1 --transcriptome=refdata-gex-GRCh38-2020-A --fastqs=/path/to/sample1/fastq/folder
cellranger count --id=sample2 --transcriptome=refdata-gex-GRCh38-2020-A --fastqs=/path/to/sample2/fastq/folder
# Repeat the above command for each sample, adjusting the sample name and fastq folder path


## python conda environment managment

# get the current python user
import getpass

username = getpass.getuser()
print("Current username:", username)

for installing packages in the right environment the terminal and the jupyter notebook environemnt should match.
for this we can use the environment managemnt from juypter notebook or the terminal. 



#### how to install packages in terminal and make them available in your jupyternotebook

#In the terminal initialize conda, create an environment, activate that environemnt 
#and make it available for jupyternotebooks in jupyterhub
# explained here https://blog.umd.edu/crstl/useful-links/how-to-setup-and-use-your-own-conda-environment-in-jupyterhub/
conda create –name myenv python=3.6

conda create –name myenv

conda activate myenv

conda install PACKAGENAME

python -m ipykernel install --user --name singlecell --display-name "Python (singlecell)

#restart the kernel or server. Open notebook. Go to Kernel --> change kernel --> select your environment


#### other option for conda env mangement

!pwd

# Initialize the terminal for using conda (maybe not necessary)
#!source ~/.conda/etc/profile.d/conda.sh

!conda init
# restart terminal --> got to control panel (stop server & start server)

!conda init
# restart terminal --> got to control panel (stop server & start server)

!pwd

# query existinmg environments
!conda info --envs

!conda deactivate

# install pakcages from notebook (not recommended) 
!conda install --yes --prefix {sys.prefix} scanpy

!conda list --name singlecell

!conda list

import scipy

# create a test environemnt if not done yet (maybe not neccessary)
!conda create --name test2 -y


!conda activate test 


!conda deactivate

## where is python executed?
import sys

print(sys.executable)

## what is my current system environment (weird)
import os

current_environment = os.path.basename(sys.prefix)
print("Current environment:", current_environment)

## what is my current conda environment (weird)
import subprocess

def get_current_environment_name():
    conda_env_command = ['conda', 'info', '--envs']
    conda_env_out = subprocess.check_output(conda_env_command, universal_newlines=True)
    
    for line in conda_env_out.strip().split('\n'):
        if '*' in line:
            return line.split()[0]

current_environment = get_current_environment_name()
print("Current environment:", current_environment)
