==> trash/04702024/10.1_prepare_data_set.ipynb <==
# Load packages

#pip install pandas

import pandas as pd

#path="result_1.csv" #Please insert correct file input name

#data=pd.read_csv(path, header=None)

!cd /home2/lumpi/Projects/single-cell_samantha/code/scripts

!pwd

# Load the counts stored from the group selection of the adata object

import pandas as pd

# Define the path to the CSV file
path = "/home2/lumpi/Projects/single-cell_samantha/data/output.csv"

# Read the CSV file, without considering any row as header initially, and using low_memory=False
#data = pd.read_csv(path, header=None, low_memory=False)
data = pd.read_csv(path, low_memory=False)

data 

# Adjust columns to the PFA format

# Manually set the first row as the header
data.columns = data.iloc[0].astype(str)

# Drop the first two rows now that headers are set
data = data.drop([0])

# Reset the index of the DataFrame
data.reset_index(drop=True, inplace=True)

data

gene_names=pd.DataFrame({'gene_names': list(data.iloc[:,0])})
expression_data=data.iloc[:,1:]

expression_data

gene_names.to_csv('/home2/lumpi/Projects/single-cell_samantha/data/gene_names.csv',index=False)
expression_data.to_csv('/home2/lumpi/Projects/single-cell_samantha/data/PFA_analysis_data.csv',index=False, header=True)

## Compare gene positions of the dataframe gene_names and expression_data

### Gene position in gene_names

gene_names

#gene_names['LINC02223']
gene_names.index[gene_names['gene_names'] == 'LINC02223'].tolist()

gene_names.iloc[468]

### Get position and average expression level in expression_data and data

##### Expression at position X of expresison data

df = expression_data.transpose()

df

df.index

for col in df.columns:
    # Try converting each column to numeric, coercing errors to NaN
    df[col] = pd.to_numeric(df[col], errors='coerce')

## Group by Index
#average_values = df.groupby(df.iloc[:, 0]).mean()
average_values = df.groupby(df.index).mean()
average_values

#"LINC02223"
#average_column_2 = average_values.iloc[:, 1]
average_column_2 = average_values.iloc[:,468]
average_column_2

## Validate smae location of gene_names and gene name
gene_names.iloc[468]

##### Validate Expression at position X of data

data.T
df = data.T
df

df.columns = df.iloc[0].astype(str)
#df = df.drop(['Label'])
df = df[1:]

df

for col in df.columns:
    # Try converting each column to numeric, coercing errors to NaN
    df[col] = pd.to_numeric(df[col], errors='coerce')

#average_values = df.groupby(df.iloc[:, 0]).mean()
average_values = df.groupby(df.index).mean()
average_values.iloc[:, 467]

average_values['LINC02223']
