==> trash/04702024/10.2_PFA_gene_selection.ipynb <==
#pip install networkx
#pip install scipy
#pip install principal-feature-analysis
#pip install pandas

from principal_feature_analysis import pfa 
import pandas as pd

!pwd

# Run PFA

# test how the pfa read the data (somehow the label row gets processed)
path="/home2/lumpi/Projects/single-cell_samantha/data/PFA_analysis_data.csv"
data = pd.read_csv(path, sep=',', header=None)
data

res,mutual_information=pfa(path="/home2/lumpi/Projects/single-cell_samantha/data/PFA_analysis_data.csv", calculate_mutual_information=True,
                           cluster_size=200, min_n_datapoints_a_bin=50)

#transfor into dataframe
mutual_information=mutual_information[0]

mutual_information

# correct the index of mutual information
mutual_information = mutual_information.drop(mutual_information.index[0])
mutual_information

mutual_information['index feature'] = mutual_information['index feature'] - 1
mutual_information

## correct res index (not sure if necessary)
#type(res)
# Delete the first element of list 'res'
del res[0]

# Subtract 1 from each element in the list
res = [x - 1 for x in res]
res[0:5]

# refine the block where combining mutual information

gene_names=pd.read_csv('/home2/lumpi/Projects/single-cell_samantha/data/gene_names.csv')
# Set index for gene_names DataFrame to match the join operation requirement
gene_names.index = range(gene_names.shape[0])  # Assuming gene_names DataFrame should match indices as given
mutual_information = mutual_information.join(gene_names, on='index feature', how='left')
mutual_information_sorted = mutual_information.sort_values(by=["mutual information"], ascending=False)

mutual_information_sorted.to_csv('/home2/lumpi/Projects/single-cell_samantha/data/gene_mutual_information.csv', index=False)

f = open("principal_features_for_model.txt", "w")
for i in res:
    f.write(str(i)+str(","))
f.close()

mutual_information_sorted.head(20)

## Of the gene List

gene_names=pd.read_csv('/home2/lumpi/Projects/single-cell_samantha/data/gene_names.csv')

gene_names.shape

## ## Validate smae location of gene_names and gene name
gene_names.iloc[468]

gene_names=pd.read_csv('/home2/lumpi/Projects/single-cell_samantha/data/gene_names.csv')

### Of the PFA file

path="/home2/lumpi/Projects/single-cell_samantha/data/PFA_analysis_data.csv"
number_sweeps=10
gene_selection=0 # validate on PFA = 0, random genes = 1 or all genes = 2


data_total=pd.read_csv(path, sep=",", header=None).transpose()

## here read correctly
data=data_total.sample(frac = 0.8)
data_test=data_total.drop(data.index)
data=data.transpose().to_numpy()
data_test=data_test.transpose().to_numpy()
data_total

data_test

df = data_total
#average_values = df.groupby(df.index).mean()
average_values = df.groupby(df.iloc[:, 0]).mean()
average_values.iloc[:, 468]  
## this position (468) is not LINC02223 but MYO10 
## assumed is when PFA tells a position the wrong gene names is assigned. 
## However the validation with this wrong assigned genne names works, indicating that the PFA alogorithm might account for the mismatch in column index)

#gene_names['LINC02223']
gene_names.index[gene_names['gene_names'] == 'LINC02223'].tolist()

gene_names=pd.read_csv('/home2/lumpi/Projects/single-cell_samantha/data/gene_names.csv')



gene_names=pd.read_csv('/home2/lumpi/Projects/single-cell_samantha/data/gene_names.csv')
gene_names=gene_names.iloc[res,:].reset_index(drop=True)

mutual_information=mutual_information.join(gene_names)

mutual_information_sorted=mutual_information.sort_values(by=["mutual information"], ascending=False)
mutual_information_sorted.to_csv('/home2/lumpi/Projects/single-cell_samantha/data/gene_mutual_information.csv', index=False)

gene_names.iloc[res,:][200:220]

import inspect 
print(inspect.getsource(pfa))

def pfa(path, number_output_functions=1, number_sweeps=1, cluster_size=50, alpha=0.01, min_n_datapoints_a_bin=500, shuffle_feature_numbers=0, frac=1, calculate_mutual_information=0, basis_log_mutual_information=2):
    # pf_ds = principal features related to output functions, pf = all principal features
    start_time=time.time()
    list_pf_ds=[]


    # The csv file's content is an m x n Matrix with m - number components of output-function = number features and n = number of data points
    # where the first number components of output-function rows contain the value of the vector-valued output function for each of the n data points
    # e.g. in case of a one-dimensional output function, the first row can be the label for each data point
    data = pd.read_csv(path, sep=',', header=None)
    for sweep in range(0,number_sweeps):
        print("Sweep number: " + str(sweep+1))
        pf_ds,pf,indices_principal_feature_values=find_relevant_principal_features(data,number_output_functions,cluster_size,alpha,min_n_datapoints_a_bin,shuffle_feature_numbers,frac)
        list_pf_ds.append(pf_ds)
        # Output the principal features related to the output function in a list where the numbers correspod to the rows of the input csv-file
        f = open("principal_features_depending_system_state"+str(sweep)+".txt", "w")
        for i in pf_ds:
            for j in i:
                f.write(str(j) + str(","))
            f.write("\n")
        f.close()
        # Output the principal features in a list where the numbers correspond to the rows of the input csv-file
        f = open("principal_features_global_indices"+str(sweep)+".txt", "w")
        for i in pf:
            for j in i:
                f.write(str(j) + str(","))
            f.write("\n")
        f.close()
        np.savetxt("global_indices_and_principal_features_state_dependency"+str(sweep)+".csv", indices_principal_feature_values,delimiter=",")
    print("Time needed for the PFA in seconds: " + str(time.time()-start_time))


    #Intersect the lists of principal features related to the output function
    #All the features corresponding to the returned subgraphs are considered in each list

    list_principal_features_depending_on_system_state_for_intersection=[]
    for i in list_pf_ds:
        intermediate_list = []
        for j in i:
            for k in j:
                if k !='*':
                    intermediate_list.append(k)
        list_principal_features_depending_on_system_state_for_intersection.append(intermediate_list)
    pf_from_intersection=list_principal_features_depending_on_system_state_for_intersection[0]
    if number_sweeps > 1:
        for i in range(1, len(list_principal_features_depending_on_system_state_for_intersection)):
            pf_from_intersection=list(set(pf_from_intersection).intersection(set(list_principal_features_depending_on_system_state_for_intersection[i])))
        f = open("principal_features_depending_system_state_intersection.txt", "w")
        for i in pf_from_intersection:
            f.write(str(i)+str(","))
        f.close()

    # Outputs a list of DataFrames where the index feature refers to the row in the csv-file.
    # The mutual information is calculated between the feature represented in the first row of the data frame and the feature referred to in the index feature column.
    # The first row is consequently the mutual information of the corresponding component of the output-function with itself
    if calculate_mutual_information==1:
        print("Calculating mutual information")
        list_data_frame_feature_mutual_information=get_mutual_information(data,number_output_functions,pf_from_intersection,min_n_datapoints_a_bin,basis_log_mutual_information)
        for i in range(0,len(list_data_frame_feature_mutual_information)):
            print(list_data_frame_feature_mutual_information[i])
        return pf_from_intersection, list_data_frame_feature_mutual_information
        
    return pf_from_intersection


print(inspect.getsource(find_relevant_principal_features))
